requirements.txt
pymongo==4.4.1
python-dotenv
streamlit==1.29.0
matplotlib==2.8.2
pandas==2.2.0